Working clients

+++ src/grocerystore.h
@@ -14,16 +14,16 @@ typedef enum {
 typedef struct {
     pthread_mutex_t mutex;
     gs_state state;
-    size_t clients_inside;
+    int clients_inside;
     pthread_cond_t exit;
     pthread_cond_t entrance;
-    size_t can_enter;
+    int can_enter;
 } grocerystore_t;
 
 grocerystore_t *grocerystore_create(size_t e);
 void grocerystore_destroy(grocerystore_t *gs);
 gs_state enter_store(grocerystore_t *gs, size_t id);
 void exit_store(grocerystore_t *gs);
-gs_state doBusiness(grocerystore_t *gs, size_t c, size_t e);
+gs_state doBusiness(grocerystore_t *gs, int c, int e);
 
 #endif //GROCERYSTORE_H
